refactor: moved book dto to nested package
