+++ hw-9/src/main/java/ru/dankoy/hw7/core/dto/book/mapper/BookMapperImpl.java
@@ -1,9 +1,9 @@
-package ru.dankoy.hw7.core.dto.mapper;
+package ru.dankoy.hw7.core.dto.book.mapper;
 
 import java.util.HashSet;
 import org.springframework.stereotype.Component;
 import ru.dankoy.hw7.core.domain.Book;
-import ru.dankoy.hw7.core.dto.BookDTO;
+import ru.dankoy.hw7.core.dto.book.BookDTO;
 
 
 @Component
