+++ hw-9/src/main/java/ru/dankoy/hw7/core/service/book/BookDtoService.java
@@ -2,7 +2,7 @@ package ru.dankoy.hw7.core.service.book;
 
 import java.util.List;
 import java.util.Optional;
-import ru.dankoy.hw7.core.dto.BookDTO;
+import ru.dankoy.hw7.core.dto.book.BookDTO;
 
 public interface BookDtoService {
